Added PointProperties to shader.

+++ libtitania-x3d/share/titania/shaders/Shaders/Types.h
@@ -38,7 +38,7 @@ struct x3d_PointPropertiesParameters
 	mediump float pointSizeMinValue;
 	mediump float pointSizeMaxValue;
 	mediump vec3  pointSizeAttenuation;
-	int   colorMode;
+	mediump int   colorMode;
 };
 
 //uniform x3d_PointPropertiesParameters x3d_PointProperties;
